+++ mmvstr.h
@@ -187,8 +187,8 @@ vstring quo$(vstring sout);
 /******* Special purpose routines for better
       memory allocation (use with caution) *******/
 
-extern int tempAllocStackTop;   /* Top of stack for tempAlloc functon */
-extern int startTempAllocStack; /* Where to start freeing temporary allocation
+extern long tempAllocStackTop;   /* Top of stack for tempAlloc functon */
+extern long startTempAllocStack; /* Where to start freeing temporary allocation
     when let() is called (normally 0, except for nested vstring functions) */
 
 /* Make string have temporary allocation to be released by next let() */
